requirements.txt
# Federal API Vault - Production Dependencies

# Core HTTP client
requests>=2.31.0

# Environment configuration
python-dotenv>=1.0.0

# Data validation (optional but recommended)
pydantic>=2.5.0

# Logging enhancements
colorlog>=6.8.0

# CLI enhancements
click>=8.1.7

# Testing (dev dependencies)
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0  # Mock HTTP responses
